fix(deps): bump mobile bridges to observability-android 0.65.0 and client SDK 5.13.3 (#720)
## Summary

- Bump the Flutter and React Native native bridges to
`com.launchdarkly:launchdarkly-observability-android:0.65.0` (latest on
Maven Central).
- Bump `observability-android` and the React Native bridge to
`com.launchdarkly:launchdarkly-android-client-sdk:5.13.3` (latest on
Maven Central).
- 0.65.0 widened `ReplayOptions.scale` from `Float` to a nullable
`Double`, so the React Native bridge stops narrowing the JS-provided
scale to `Float` and its unit tests assert `Double` values. The Flutter
bridge already passed a `Double`.

## Test plan

- [x] `./gradlew test` in `sdk/@launchdarkly/observability-android`
- [x] `./gradlew :app:compileDebugKotlin` in
`sdk/@launchdarkly/react-native-ld-session-replay/example/android`
- [x] `./gradlew
:launchdarkly_session-replay-react-native:testDebugUnitTest --tests
'com.sessionreplayreactnative.SessionReplayClientAdapterTest'`
- [ ] CI: Android observability, Android e2e, Flutter, and React Native
session replay workflows

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Overview**
> Bumps **Flutter**, **React Native session replay**, and
**observability-android** Gradle deps to
`launchdarkly-observability-android:0.65.0` and
`launchdarkly-android-client-sdk:5.13.3`.
> 
> The React Native bridge passes JS `scale` through as a **Double** into
`ReplayOptions` (no `toFloat()`), matching observability **0.65.0**
where `ReplayOptions.scale` widened from `Float` to nullable `Double`.
Unit tests assert `Double` scale values accordingly. Flutter only
updates the observability artifact version and a comment about the
client SDK pin.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
432cf5ec00899d6fe27beadde1a4b3c16705aafb. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ sdk/@launchdarkly/react-native-ld-session-replay/android/src/main/java/com/sessionreplayreactnative/SessionReplayClientAdapter.kt
@@ -299,7 +299,7 @@ internal class SessionReplayClientAdapter private constructor() {
             enabled = isEnabled,
             sampleRate = sampleRate,
             frameRate = frameRate,
-            scale = replayScale.toFloat(),
+            scale = replayScale,
             privacyProfile = PrivacyProfile(
                 maskTextInputs = maskTextInputs,
                 maskWebViews = maskWebViews,
